Implemented - DeckData + CS Card Assignment
This commit adds the DeckData static script which keeps track of the current games cards and if they've been assigned or not. The CardData script has also been added, this keep track of the objs assigned card name and shown status.

The ATCManagement script is the bulk of this commit as it implements the actual card assignment for Clock Solitaire and makes sure each assignment is random.

Next, start work on allowing the player to begin the game by showing their first middle card.