Implemented - cardData.cardNum Field
This commit adds functionality to split the card number from it's name and set the new cardNum field. This will allow me to start to differentiate which pile the player needs to put the active card in when dragging.

Next, continue to implement the functionality for showing the player where to place the active card.