Implemented/Fixed - Proper Full Game + Every Card Interactable
This commit implements the full Clock Solitarie game (end screen is still missing), and also fixed the issue where every card was interactable even if not the active card.

Next, make sure the game has an actual end screen (possibly with a score calculated on how many cards were placed before losing) and also prevent the player from getting stuck if a full pile is made, except for Kings as this ends the game.