Implemented - Percentage Calculation
This commit implements functionality for the calculation of the final game percentage that is shown on the end screen. This percentage lets the player know how close they were to beating Clock Solitaire.

Next, implement a button that shows a page containing the rules in the top right corner of the UI.