Adjusted - Clock Solitare Card Pile Positions
This commit makes the final adjustments to the Clock Solitaire card pile positions so they fit much better on the screen. UI changes need to be made at some point so the the card piles adjust in size to fit different screen sizes.

Next, begin work on the New Market game.