Implemented - Card Hand Ordering
This commit adds a merged main player hand which serves as a simple button that will eventually allow the player to see their hand and select a card to play. It also implements functionality to sort each players hand (and the dummy hand) into suits and rank.

Next, implement functionality to allow the player to see their hand when their hand is clicked.