Implemented - Card Hover
This commit implements functionality to allow the player to hover over each one of their cards giving them the impresion that they have selected that one.

Next, allow the player to follow up the hover with a click to remove the card from their hand and place it in the pile.