Implemented - BasicHover Script For Title Cards
This commit adds the BasicHover script which has the games title appear on the title card when the player hovers over it. This makes it possible to tell which game is which without having text above the title cards as this didn't look right.

Next, go back to programming New Market.