Implemented - Player Starting Tokens
This commit introduces the players starting tokens and a way to access them. A visual representation of them has also been placed next to each players hand.

Next, implement a way to have each CPU player auto place a token at the beginning of play and pick a horse for their 2nd token.