Implemented - Starting Token Placement (Player)
This commit implements the ability for the player to place their own tokens down in the middle and on one of the four horses.

Next, create a static script which keeps track of the variables changing and make sure the NewMarketManagement Script can keep track of what's currently happening in the game.