Implemented - Improved CPU Card Laying
This commit improves upon the CPU card laying once again by removing the beforeLay images and actually adding an outline to the layed card with a time gap in between each layed card so that the player can see which card has been layed.

Next, prevent the player from being able to select any card and instead have them only be able to select the lowest of a suit.