Implemented - Next Card Only Enabled
This commit prevents the player from being able to just lay their lowest black, it now allows them to only lay their lowest black on the first turn and then only the next card in the suit if they have it. Otherwise they can't lay any cards.

Next, implement functionality for the game to continue (by switching suits & colour) when a suit is complete or can't be continued as the next card is in the dummy hand.