Implemented - ExtractSuitAndRank() Method
This commit implements the ExtractSuitAndRank() method which is used to prevent  code repetition within the ShowPlayerHand() method. It extract the suit and card rank from the given card in the form of a List<string>.

Next, implement functionality for the game to continue (by switching suits & colour) when a suit is complete or can't be continued as the next card is in the dummy hand.