Implemented - Check For ChangeSuit Call
This commit implements a place for the ShowPlayerHand method to check if the player should be placing a new card down in a different suit. This only occurs when the CPUs don't have the next card or the suit has been finished for the player.

Next, implement the actual functionality for the ChangeSuit method.