Implemented - Player Suit Switching
This commit finishes implementing functionality for the players suit switching, after a suit has finished or hit a dead end the player will now be able to place down another card of a different colour suit IF they were the last to lay.

Next, implement this same functionality but for the CPUs.