Implemented - CPU Suit Switching
This commit implements the functionality for the CPU suit switching, it allows each CPU player to react to when a suit has hit a dead end and they need to place down a new card of the opposite colour.

Next, fix the issue where if a card has already been placed the CPU can't recognise this. This is likely due to the CPU only checking the dummyHand.