Implemented - LayedCards Script Methods
This commit updates the LayedCards Script with a new UpdateStaticCardData method which just contains the original code from that script but placed into a method instead of Update.

I have also implemented a check for when a horse card has been layed and started the CollectHorse method.

Next, implement functionality for the CollectHorse() method.