Fixed - Player Suit Change Occurs Early
This commit fixes an issue where the player would be able to change suit if a CPU didn't immediatly place done the next card, now the ShowHand script checks if the NMStaticData.shouldWait variable has been trigged first.

If it has, don't allow a suit change.

Next, implement functionality to allow the CPU to change suit if the suit has been completed (the Ace of the suit has been layed).