Implemented - Change Suit at End of Suit Functionality
This commit first moves the change suit CPUBehaviour code into a ChangeSuit() method to prevent code reuse and then implements the nessecary checks to make sure the CPUs can change suit once the Ace of that suit has been layed as this is the end of the suit.

Next, prevent the CPU & player from getting stuck if they can't switch to a different colour suit. Instead allow them to place one of the same colour just layed.