Implemented - CPU Player Drop Outs
This commit implements functionality for when the CPUs run out of tokens to buy into a round. Once they don't have enough to buy-in again each remaining player is given their share of the next rounds cards and the visual representation of their hand is faded out.

Next, implement a way for the game to end with a GAME OVER if the player runs out of tokens to continue buying in.